refactor: move update cart service to API file
